YP-26 fix: fix routes error
поправил иерархию роутов, поставил новую стору как основную, добавил навигационную панель на главную страницу
