+++ packages/client/src/Components/App/App.scss
@@ -18,6 +18,6 @@
 
 .game-field__canvas {
   width: calc(100% - 300px);
-  height: 90vh;
+  height: calc(90vh - 64px);
   border: 1px solid gray;
 }
