+++ packages/client/src/Constants/Routes.ts
@@ -1,0 +1,8 @@
+export const Routes = {
+  Main: '/',
+  SignUp: 'sign-up',
+  SignIn: 'sign-in',
+  Forum: 'forum',
+  LeaderBord: 'leader-bord',
+  Profile: 'profile',
+} as const;
